+++ booktrading/src/bookTrading/seller/BookSeller.java
@@ -4,6 +4,6 @@ import java.util.Date;
 
 public interface BookSeller {
 
-	public void sell(String bookTitle, int initPrice, int minPrice, Date deadline);
+	public void sell(String bookTitle, double initPrice, double minPrice, Date deadline);
 	
 }
